% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalpha.R
\name{kalpha}
\alias{kalpha}
\title{Compute Krippendorff's Alpha}
\usage{
kalpha(dt, unit, measurement, level)
}
\arguments{
\item{dt}{`data.table` containing the reliability data in long format.}

\item{unit}{Name of the column containing the unit ID}

\item{measurement}{Name of the column containing the measurements, one per
judge}

\item{level}{c('binary', 'nominal'): type of oberverment data.}
}
\value{
\item{alpha}{Krippendorff's Alpha reliability index}
\item{De}{Expected disagreement}
\item{Do}{Overall observed disagreement accross all units}
\item{by_unit}{Dataframe with one line per unit and columns}
\item{unit}{Unit}
\item{mu}{Number of observations in that unit}
\item{Do}{Observed disagreement within this unit}
}
\description{
This function implements the computation of  Krippendorff's Alpha as per
https://repository.upenn.edu/cgi/viewcontent.cgi?article=1043&context=asc_papers # nolint
}
\details{
It is designed to be space efficient for sparse oberverments, and as thus
does not take as input a reliability matrix, but a long-format data.table

If a tibble or a non-data.table dataframe is passed as input, this function
will still work! It will silently create a data.table copy of your dataframe
(or tibble).
TODO(jucor): cite properly
}
