% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwebreliability5.R
\docType{data}
\name{mwebreliability5}
\alias{mwebreliability5}
\title{Example data from Krippendorff 2013}
\format{
A `data.table` with 41 rows and 3 variables:
\describe{
\item{unit}{Unit observed}
\item{observer}{Name of the observer}
\item{measurement}{Measurement from the observer, between 1 and 4}
}
}
\usage{
mwebreliability5
}
\description{
Example reliablility data, with multiple observers and missing data,
from Krippendorf (2013, page 4), 4 observers, 12 units. This example can be
interpreted either as nominal, ordinal, or interval data
}
\details{
It is cast to_long_form,
with missing observations removed from the dataset.
See \code{\link{mwebreliability5.wide}} for the wide version ontaining NA.

TODO(jucor): Find how to cite Hayes 2013 properly with @references
}
\seealso{
\code{\link{mwebreliability5.wide}}
}
\keyword{datasets}
