% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_presence_wide.R
\docType{data}
\name{news_presence_wide}
\alias{news_presence_wide}
\title{Two observers, binary reliability data, section 11.3.1 from Krippendorff 2004}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{article}{unit observed}
\item{jon}{Jon's binary judgement}
\item{han}{Han's binary judgement}
}
}
\usage{
news_presence_wide
}
\description{
This is the wide version of \code{\link{news_presence}}.
}
\seealso{
\code{\link{news_presence}}
}
\keyword{datasets}
