% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_tone_wide.R
\docType{data}
\name{news_tone_wide}
\alias{news_tone_wide}
\title{News coverage reliability data set, Table 1 from Hayes & Krippendorff 2007}
\format{
A data frame with 40 rows and 6 variables:
\describe{
\item{unit}{unit observed}
\item{obs1}{ordinal judgement from observer 1}
\item{obs2}{ordinal judgement from observer 2}
\item{obs3}{ordinal judgement from observer 3}
\item{obs4}{ordinal judgement from observer 4}
\item{obs5}{ordinal judgement from observer 5}
}
}
\usage{
news_tone_wide
}
\description{
Wide format of the data in \code{\link{news_tone}}.
}
\seealso{
\code{\link{news_tone}}
}
\keyword{datasets}
