% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogacy.R
\name{surrogacy}
\alias{surrogacy}
\title{Compute Krippendorff's Surrogacy}
\usage{
surrogacy(
  coders,
  standard,
  unit_from = "unit",
  measurement_from = "measurement",
  frequency_from = NULL,
  return_by_unit = FALSE
)
}
\arguments{
\item{coders}{`data.table` containing the coders.}

\item{standard}{`data.table` containing the standard.}

\item{unit_from}{Name of the column containing the unit ID}

\item{measurement_from}{Name of the column containing the measurements}

\item{frequency_from}{(Optional) Name of the column containing the
frequencies, *if* the data is in the "aggregated" form described above.}

\item{return_by_unit}{(default FALSE) If TRUE, return a data.table of
per-unit contributions.}
}
\value{
\item{surrogacy}{Krippendorff's Alpha surrogacy index}
\item{De}{Expected disagreement}
\item{Do}{Overall observed disagreement accross all units}
\item{by_unit}{(Only if return_by_unit = TRUE) Per-unit contributions}
}
\description{
Similar input and warnings as [replicability()].
TODO(julien): figure out how to merge both functions in the same help page.
}
