% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalpha.R
\name{to_long_form}
\alias{to_long_form}
\title{Turn an observation data.frame into a long data.table}
\usage{
to_long_form(dt, unit, observers, measurements)
}
\arguments{
\item{dt}{`data.table` containing the reliability data in wide format}

\item{unit}{Name of the column containing the unit ID}

\item{observers}{List of names of the columns containing the oberverments,
one per observer}

\item{measurements}{Name of the new column containing the measurements}
}
\value{
A long-form melted data.table with the same unit column and two new
  columns "observer" and measurements and as many rows as units.
}
\description{
Turn an observation data.frame into a long data.table
}
